feat: add margins according to size variant
