+++ packages/x-tailwindcss/src/x-tailwind-plugin/components/suggestion-group/ghost.ts
@@ -13,9 +13,24 @@ export function suggestionGroupGhost(helpers: TailwindHelpers) {
     ghost: {
       backgroundColor: 'transparent',
       '&:hover': {
-        backgroundColor: theme('colors.neutral.10')
+        backgroundColor: theme('colors.neutral.10'),
+        '.suggestion': {
+          textDecoration: 'none'
+        }
       },
-      gap: theme('spacing.4')
+      gap: theme('spacing.4'),
+      '> *:first-child:not(.suggestion-group-button)': {
+        marginInlineStart: theme('spacing.8')
+      },
+      '> *:last-child:not(.suggestion-group-button)': {
+        marginInlineEnd: theme('spacing.8')
+      },
+      '&.suggestion-group-lg > *:first-child:not(.suggestion-group-button)': {
+        marginInlineStart: theme('spacing.12')
+      },
+      '&.suggestion-group-lg > *:last-child:not(.suggestion-group-button)': {
+        marginInlineEnd: theme('spacing.12')
+      }
     }
   };
 }
